requirements.txt
# Core Bot Dependencies
python-telegram-bot==20.7
apscheduler==3.10.4
python-dotenv==1.0.0

# HTTP and API
requests==2.31.0
urllib3==2.0.7

# UI and Display
rich==13.7.0

# System Monitoring
psutil==5.9.6

# Date and Time
pytz==2023.3
tzdata==2023.3

# JSON and Data Processing
jsonschema==4.19.2

# Logging and Debugging
colorlog==6.8.0

# Security and Environment
cryptography==41.0.7

# Optional: For database support (if you want to add persistence)
# sqlite3 (built-in with Python)

# Optional: For advanced monitoring
# prometheus_client==0.19.0

# Development and Testing
# pytest==7.4.3
# pytest-asyncio==0.21.1